Simplify type_traits guards
We have to support building in environments that
don't support <type_traits> at all, or only support
some of the features we need. Complicating matters
is the fact that not all features have standard test
macros, nor can we rely on the code being built with
CMake (which would allow us to detect compiler
support for specific features at build-config time).
Instead we'll rely on the __cplusplus macro to detect
C++11.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>