Give cgroup v1 mount point macro a better name
The mount point is not v1-specific, v2 can also be mounted there. The
mount point is also not guaranteed to be at /sys/fs/cgroup, it just happens
to be there on most distros by convention. Therefore, rename the macro
to OMR_CGROUP_DEFAULT_MOUNT_POINT to avoid misleading readers.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>